Move Phi4 oracle helpers out of main
